requirements.txt
install==1.3.5
joblib==1.2.0
numpy==1.24.3
progress==1.6
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
python-utils==3.5.2
scikit-learn==1.2.2
scipy==1.10.1
TgCrypto==1.2.5
threadpoolctl==3.1.0
